pyproject.toml
[project]
name = "animvideo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pillow>=11.3.0",
    "snakeviz>=2.2.2",
]
